Beginning of adding S2oE Support
Added in a timer to check if new packets received on listened port.

Currently the "Receive" button acts as a toggle.. which I should change to be visibly actively receiving/not receiving because it's not obvious.

I tested this with my S2oE client I use for testing networking setups for westrace gear and it receives packets fine.

I also added in the Send area the foundations for sending S2oE packets to Westrace.. I'm not sure this is needed functionality for Logic Navigator but is useful for debugging purposes during developement.